Move payment channel storage API into separate module.
